preppal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
faster-whisper==1.2.0
openai-whisper
gtts==2.4.0
